Add Thrift statistics requests needed by transactional tables
ANALYZE applied to a Hive transactional table fails because the
Hive metastore statistics update requests used by Trino don't
contain the writeId and validWriteIds string.  This commit adds
the additional structures and metastore methods needed to update
statistics for transactional tables.

This commit also makes a couple of changes to the generate-sources.sh
script that appear to be required.